Delete event complete, event related components moved to better reflect structure
